requirements.txt
numpy==1.16.4
Keras==2.2.4
tqdm==4.23.2
pycuber==0.2.2
tensorflow-gpu
h5py==2.10.0
